[#286] Upgrade to TS4, fix type errors
