+++ package.json
@@ -100,7 +100,7 @@
     "svgson": "^4.0.0",
     "ts-jest": "^24",
     "ts-loader": "^8.0.7",
-    "typescript": "^3.9.5"
+    "typescript": "4"
   },
   "resolutions": {
     "@types/react": "^16.14.2"
