+++ src/ebay-tabs/index.ts
@@ -1,4 +1,4 @@
 export { default as EbayTabs } from './tabs'
 export { default as EbayTab } from './tab'
 export { default as EbayTabPanel } from './tab-panel'
-export { Activation, Size } from './types'
+export type { Activation, Size, TabsProps, OnSelectProps } from './types'
